+++ amdsmi_cli/amdsmi_logger.py
@@ -170,6 +170,18 @@ class AMDSMILogger():
                 table_values += string_value.ljust(18)
             elif key == "RW":
                 table_values += string_value.ljust(57)
+            elif key in ('pviol', 'tviol'):
+                table_values += string_value.rjust(7)
+            elif key == "tviol_active":
+                table_values += string_value.rjust(14)
+            elif key == "phot_tviol":
+                table_values += string_value.rjust(12)
+            elif key == "vr_tviol":
+                table_values += string_value.rjust(10)
+            elif key == "hbm_tviol":
+                table_values += string_value.rjust(11)
+            elif key == "gfx_clkviol":
+                table_values += string_value.rjust(13)
             elif key == "process_list":
                 #Add an additional padding between the first instance of GPU and NAME
                 table_values += '  '